requirements.txt
accelerate==1.8.1
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
appnope==0.1.4
asttokens==3.0.0
attrs==25.3.0
beautifulsoup4==4.13.4
cbor==1.0.0
certifi==2025.6.15
charset-normalizer==3.4.2
coloredlogs==15.0.1
comm==0.2.2
dataclasses-json==0.6.7
datasets==3.6.0
debugpy==1.8.14
decorator==5.2.1
dill==0.3.8
executing==2.2.0
fastembed==0.7.1
filelock==3.18.0
flagembedding==1.3.5
flatbuffers==25.2.10
frozenlist==1.7.0
fsspec==2025.3.0
grpcio==1.73.0
h11==0.16.0
h2==4.2.0
hf-xet==1.1.5
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.1
huggingface-hub==0.33.0
humanfriendly==10.0
hyperframe==6.1.0
idna==3.10
ijson==3.4.0
inscriptis==2.6.0
ipykernel==6.29.5
ipython==9.3.0
ipython-pygments-lexers==1.1.1
ir-datasets==0.5.11
jedi==0.19.2
jinja2==3.1.6
joblib==1.5.1
jsonpatch==1.33
jsonpointer==3.0.0
jupyter-client==8.6.3
jupyter-core==5.8.1
kiwipiepy==0.21.0
kiwipiepy-model==0.21.0
langchain==0.3.26
langchain-community==0.3.26
langchain-core==0.3.66
langchain-text-splitters==0.3.8
langsmith==0.4.1
loguru==0.7.3
lxml==5.4.0
lz4==4.4.4
markupsafe==3.0.2
marshmallow==3.26.1
matplotlib-inline==0.1.7
mmh3==5.1.0
mpmath==1.3.0
multidict==6.5.1
multiprocess==0.70.16
mypy-extensions==1.1.0
nest-asyncio==1.6.0
networkx==3.5
numpy==2.3.1
onnxruntime==1.22.0
orjson==3.10.18
packaging==24.2
pandas==2.3.0
parso==0.8.4
peft==0.15.2
pexpect==4.9.0
pillow==11.2.1
platformdirs==4.3.8
portalocker==2.10.1
prompt-toolkit==3.0.51
propcache==0.3.2
protobuf==6.31.1
psutil==7.0.0
ptyprocess==0.7.0
pure-eval==0.2.3
py-rust-stemmers==0.1.5
pyarrow==20.0.0
pydantic==2.11.7
pydantic-core==2.33.2
pydantic-settings==2.10.1
pygments==2.19.2
pymupdf==1.26.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
pyyaml==6.0.2
pyzmq==27.0.0
qdrant-client==1.14.3
regex==2024.11.6
requests==2.32.4
requests-toolbelt==1.0.0
safetensors==0.5.3
scikit-learn==1.7.0
scipy==1.16.0
sentence-transformers==4.1.0
sentencepiece==0.2.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
sqlalchemy==2.0.41
stack-data==0.6.3
sympy==1.14.0
tenacity==9.1.2
threadpoolctl==3.6.0
tokenizers==0.21.1
torch==2.7.1
tornado==6.5.1
tqdm==4.67.1
traitlets==5.14.3
transformers==4.52.4
trec-car-tools==2.6
typing-extensions==4.14.0
typing-inspect==0.9.0
typing-inspection==0.4.1
tzdata==2025.2
unlzw3==0.2.3
urllib3==2.5.0
warc3-wet==0.2.5
warc3-wet-clueweb09==0.2.5
wcwidth==0.2.13
xxhash==3.5.0
yarl==1.20.1
zlib-state==0.1.9
zstandard==0.23.0
